files/main: Make files.log analyzer order predictable
Introducing new packet analyzers may change the underlying enum
value of File analyzers. This has been observed twice and generally
caused some amount of confusion due re-ordering of sets in baselines:

    https://github.com/zeek/zeek/pull/2340
    https://github.com/zeek/zeek/pull/2770

Make the order in which File analyzers are added to fa_file object
predictable by switching to a vector instead of set and pre-sorting.
Additionally, use &ordered on the analyzers field.

The sorted vector is used as we can not easily change the existing
register_for_mime_type functions taking a set.